# cmd/swag/docs/swagger.yaml
basePath: /
definitions:
  dto.Candidate:
    properties:
      content:
        $ref: '#/definitions/dto.Content'
      finishMessage:
        type: string
      finishReason:
        type: string
      index:
        type: integer
    type: object
  dto.ChatCompletionMessage:
    properties:
      content:
        type: string
      role:
        type: string
    type: object
  dto.ChatCompletionRequest:
    type: object
  dto.ChatCompletionResponse:
    properties:
      choices:
        items:
          $ref: '#/definitions/dto.Choice'
        type: array
      created:
        type: integer
      id:
        type: string
      model:
        type: string
      object:
        type: string
      usage:
        $ref: '#/definitions/models.Usage'
    type: object
  dto.ChatCompletionResponseMessage:
    properties:
      content:
        type: string
      role:
        type: string
      tool_calls:
        items:
          $ref: '#/definitions/dto.ChatCompletionToolCall'
        type: array
    type: object
  dto.ChatCompletionToolCall:
    properties:
      function:
        $ref: '#/definitions/dto.ChatCompletionToolCallFunction'
      id:
        type: string
      type:
        type: string
    type: object
  dto.ChatCompletionToolCallFunction:
    properties:
      arguments:
        type: string
      name:
        type: string
    type: object
  dto.Choice:
    properties:
      finish_reason:
        type: string
      index:
        type: integer
      message:
        $ref: '#/definitions/dto.ChatCompletionResponseMessage'
    type: object
  dto.ConfigContent:
    properties:
      id:
        description: for tool_use
        type: string
      input:
        additionalProperties: true
        description: for tool_use
        type: object
      name:
        description: for tool_use
        type: string
      text:
        type: string
      type:
        description: '"text" or "tool_use"'
        type: string
    type: object
  dto.Content:
    type: object
  dto.DeepResearchRequest:
    properties:
      language:
        description: Language for the response (optional, e.g. "en", "vi")
        type: string
      max_sources:
        description: MaxSources maximum number of web sources to consult (optional)
        type: integer
      model:
        description: Model to use (optional, defaults to best available)
        type: string
      query:
        description: Query is the research topic or question
        type: string
    type: object
  dto.DeepResearchResponse:
    properties:
      completed_at:
        description: CompletedAt timestamp (Unix seconds), 0 if not completed
        type: integer
      created_at:
        description: CreatedAt timestamp (Unix seconds)
        type: integer
      duration_ms:
        description: DurationMs total research duration in milliseconds
        type: integer
      error:
        description: Error message if status is "failed"
        type: string
      id:
        description: ID of this research session
        type: string
      model:
        description: Model used
        type: string
      query:
        description: Query that was researched
        type: string
      sources:
        description: Sources used in research
        items:
          $ref: '#/definitions/dto.ResearchSource'
        type: array
      status:
        description: 'Status: "in_progress", "completed", "failed"'
        type: string
      steps:
        description: Steps of research performed
        items:
          $ref: '#/definitions/dto.ResearchStep'
        type: array
      summary:
        description: Summary is the synthesized research report
        type: string
    type: object
  dto.GeminiGenerateRequest:
    type: object
  dto.GeminiGenerateResponse:
    properties:
      candidates:
        items:
          $ref: '#/definitions/dto.Candidate'
        type: array
      usageMetadata:
        $ref: '#/definitions/dto.UsageMetadata'
    type: object
  dto.GeminiModel:
    properties:
      description:
        type: string
      displayName:
        type: string
      name:
        type: string
      supportedGenerationMethods:
        items:
          type: string
        type: array
      version:
        type: string
    type: object
  dto.GeminiModelsResponse:
    properties:
      models:
        items:
          $ref: '#/definitions/dto.GeminiModel'
        type: array
    type: object
  dto.InlineData:
    properties:
      data:
        type: string
      mimeType:
        type: string
    type: object
  dto.MessageRequest:
    type: object
  dto.MessageResponse:
    properties:
      content:
        items:
          $ref: '#/definitions/dto.ConfigContent'
        type: array
      id:
        type: string
      model:
        type: string
      role:
        description: '"assistant"'
        type: string
      stop_reason:
        type: string
      type:
        description: '"message"'
        type: string
      usage:
        $ref: '#/definitions/models.Usage'
    type: object
  dto.ResearchSource:
    properties:
      domain:
        description: Domain of the source
        type: string
      snippet:
        description: Snippet relevant excerpt from the source
        type: string
      title:
        description: Title of the source page
        type: string
      url:
        description: URL of the source
        type: string
    type: object
  dto.ResearchStep:
    properties:
      description:
        description: Description of what this step did
        type: string
      query:
        description: Query used in this step (for search steps)
        type: string
      result:
        description: Result summary of this step
        type: string
      step_number:
        description: StepNumber ordinal step number
        type: integer
      type:
        description: 'Type of step: "search", "read", "synthesize", "outline", "write"'
        type: string
    type: object
  dto.UsageMetadata:
    properties:
      candidatesTokenCount:
        type: integer
      promptTokenCount:
        type: integer
      totalTokenCount:
        type: integer
    type: object
  models.Message:
    properties:
      content:
        type: string
      role:
        type: string
    type: object
  models.ModelData:
    properties:
      created:
        type: integer
      created_at:
        type: integer
      display_name:
        type: string
      id:
        type: string
      object:
        type: string
      owned_by:
        type: string
      type:
        type: string
    type: object
  models.ModelListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.ModelData'
        type: array
      object:
        type: string
    type: object
  models.Usage:
    properties:
      completion_tokens:
        type: integer
      input_tokens:
        type: integer
      output_tokens:
        type: integer
      prompt_tokens:
        type: integer
      total_tokens:
        type: integer
    type: object
host: localhost:4981
info:
  contact: {}
  description: ✨Reverse-engineered API for Gemini web app. It can be used as a genuine
    API key from OpenAI, Gemini, and Claude.
  title: Gemini Web To API
  version: "1.0"
paths:
  /claude/v1/messages:
    post:
      consumes:
      - application/json
      description: Sends a message to the Claude model. Supports both standard JSON
        response and streaming (SSE) response.
      parameters:
      - description: Message Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.MessageRequest'
      produces:
      - application/json
      - text/event-stream
      responses:
        "200":
          description: SSE stream of dto.StreamEvent JSON objects
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Send Message (Claude)
      tags:
      - Claude
  /claude/v1/messages/count_tokens:
    post:
      consumes:
      - application/json
      description: Estimates the number of tokens for a request
      parameters:
      - description: Message Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.MessageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Count Tokens (Claude)
      tags:
      - Claude
  /claude/v1/models:
    get:
      consumes:
      - application/json
      description: Returns a list of available Claude models
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: List Claude Models
      tags:
      - Claude
  /claude/v1/models/{model_id}:
    get:
      consumes:
      - application/json
      description: Get details of a specific Claude model
      parameters:
      - description: Model ID
        in: path
        name: model_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get Claude Model
      tags:
      - Claude
  /gemini/v1beta/deepresearch:
    post:
      consumes:
      - application/json
      description: Performs deep research on a topic using Gemini.
      parameters:
      - description: Deep Research Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.DeepResearchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DeepResearchResponse'
      summary: Deep Research (synchronous)
      tags:
      - Gemini
  /gemini/v1beta/models:
    get:
      consumes:
      - application/json
      description: Returns a list of models supported by the Gemini API
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GeminiModelsResponse'
      summary: List Gemini Models
      tags:
      - Gemini
  /gemini/v1beta/models/{model}:generateContent:
    post:
      consumes:
      - application/json
      description: Generates content using the Gemini model
      parameters:
      - description: Model ID
        in: path
        name: model
        required: true
        type: string
      - description: Generate Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.GeminiGenerateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GeminiGenerateResponse'
      summary: Generate Content (Gemini)
      tags:
      - Gemini
  /gemini/v1beta/models/{model}:streamGenerateContent:
    post:
      consumes:
      - application/json
      description: Streams generated content using the Gemini model
      parameters:
      - description: Model ID
        in: path
        name: model
        required: true
        type: string
      - description: Generate Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.GeminiGenerateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Chunked JSON response
          schema:
            type: string
      summary: Stream Generate Content (Gemini)
      tags:
      - Gemini
  /health:
    get:
      description: Returns the health status of the service. Used by Docker/K8s/cloud
        platforms to monitor the service.
      produces:
      - application/json
      responses:
        "200":
          description: Service is healthy
          schema:
            properties:
              service:
                type: string
              status:
                type: string
            type: object
      summary: Health check
      tags:
      - System
  /openai/v1/chat/completions:
    post:
      consumes:
      - application/json
      description: Generates a completion for the chat message. Supports both standard
        JSON and streaming (SSE) response.
      parameters:
      - description: Chat Completion Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ChatCompletionRequest'
      produces:
      - application/json
      - text/event-stream
      responses:
        "200":
          description: SSE stream of dto.ChatCompletionChunk JSON objects
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Chat Completions (OpenAI)
      tags:
      - OpenAI
  /openai/v1/models:
    get:
      consumes:
      - application/json
      description: Returns a list of models supported by the OpenAI-compatible API
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ModelListResponse'
      summary: List OpenAI Models
      tags:
      - OpenAI
swagger: "2.0"